upload: move hocs into separate files
