+++ client/data/media/use-add-media.js
@@ -1,4 +1,3 @@
-import { createHigherOrderComponent } from '@wordpress/compose';
 import { useCallback } from 'react';
 import { getFileUploader } from 'calypso/lib/media/utils';
 import { useUploadMediaMutation } from './use-upload-media-mutation';
@@ -13,11 +12,3 @@ export const useAddMedia = () => {
 	);
 	return addMedia;
 };
-
-export const withAddMedia = createHigherOrderComponent(
-	( Wrapped ) => ( props ) => {
-		const addMedia = useAddMedia();
-		return <Wrapped { ...props } addMedia={ addMedia } />;
-	},
-	'WithAddMedia'
-);
