+++ client/data/media/with-add-external-media.js
@@ -1,0 +1,10 @@
+import { createHigherOrderComponent } from '@wordpress/compose';
+import { useAddExternalMedia } from './use-add-external-media';
+
+export const withAddExternalMedia = createHigherOrderComponent(
+	( Wrapped ) => ( props ) => {
+		const addExternalMedia = useAddExternalMedia();
+		return <Wrapped { ...props } addExternalMedia={ addExternalMedia } />;
+	},
+	'WithAddExternalMedia'
+);
