+++ client/data/media/with-add-media.js
@@ -1,0 +1,10 @@
+import { createHigherOrderComponent } from '@wordpress/compose';
+import { useAddMedia } from './use-add-media';
+
+export const withAddMedia = createHigherOrderComponent(
+	( Wrapped ) => ( props ) => {
+		const addMedia = useAddMedia();
+		return <Wrapped { ...props } addMedia={ addMedia } />;
+	},
+	'WithAddMedia'
+);
