+++ client/data/media/with-upload-site-icon.js
@@ -1,0 +1,10 @@
+import { createHigherOrderComponent } from '@wordpress/compose';
+import { useUploadSiteIcon } from './use-upload-site-icon';
+
+export const withUploadSiteIcon = createHigherOrderComponent(
+	( Wrapped ) => ( props ) => {
+		const uploadSiteIcon = useUploadSiteIcon();
+		return <Wrapped { ...props } uploadSiteIcon={ uploadSiteIcon } />;
+	},
+	'withUploadSiteIcon'
+);
