+++ client/my-sites/site-settings/podcast-cover-image-setting/index.jsx
@@ -11,7 +11,7 @@ import FormFieldset from 'calypso/components/forms/form-fieldset';
 import FormLabel from 'calypso/components/forms/form-label';
 import Image from 'calypso/components/image';
 import Spinner from 'calypso/components/spinner';
-import { withAddMedia } from 'calypso/data/media/use-add-media';
+import { withAddMedia } from 'calypso/data/media/with-add-media';
 import { createTransientMediaId } from 'calypso/lib/media/utils';
 import resizeImageUrl from 'calypso/lib/resize-image-url';
 import EditorMediaModalDialog from 'calypso/post-editor/media-modal/dialog';
@@ -26,10 +26,6 @@ import getMediaItem from 'calypso/state/media/thunks/get-media-item';
 import getMediaLibrarySelectedItems from 'calypso/state/selectors/get-media-library-selected-items';
 import { ModalViews } from 'calypso/state/ui/media-modal/constants';
 import { getSelectedSiteId, getSelectedSite } from 'calypso/state/ui/selectors';
-
-/**
- * Debug
- */
 
 import './style.scss';
