+++ client/post-editor/media-modal/index.jsx
@@ -6,8 +6,8 @@ import { connect } from 'react-redux';
 import ImageEditor from 'calypso/blocks/image-editor';
 import VideoEditor from 'calypso/blocks/video-editor';
 import CloseOnEscape from 'calypso/components/close-on-escape';
-import { withAddExternalMedia } from 'calypso/data/media/use-add-external-media';
 import { withEditMedia } from 'calypso/data/media/use-edit-media-mutation';
+import { withAddExternalMedia } from 'calypso/data/media/with-add-external-media';
 import { withDeleteMedia } from 'calypso/data/media/with-delete-media';
 import accept from 'calypso/lib/accept';
 import { bumpStat as mcBumpStat } from 'calypso/lib/analytics/mc';
